Trigger a read if data was put in the serial port just after the asynchronous read returned but before the comm mask was reset to wait for an RX_CHAR.
See codeplex item #1812, after applying patch #17708 and fixing it so all test cases pass.

After SetCommMask(maskReadPending), then getting the event, calling GetOverlappedResult(), if a byte occurs in the buffer, but before we set the CommMask, I would expect the driver to provide a CommEvent on RX_CHAR even if there was a byte already present. But it looks like not all drivers do this. So now, when we get back into the main loop, we see that readPending is false and we check if there are any bytes. If so, then we indicate data is available, and it will be read on the next iteration.