Fixed a crashing bug where calling Dispose on the SerialPortStream threw an ObjectDisposedException.
Applied patch #17665 from user "dhsmith" with minor changes.

Calling Dispose on the SerialPortStream threw an ObjectDisposedException because the call to m_EventProcessing.Reset() occurred after m_EventProcessing had been disposed. To repro this issue, hook the ErrorReceived event to a method that calls Dispose within a multithreaded environment.

I also simplified the locking code for the event processing. I didn't see a need to check for IsDisposed after every event is fired, as the 'meat' of this code does not rely on any external mutable variables, and those variables need only be checked at the tail of the loop. With this done, there's no need to wait for the event loop to stop during the Dispose method, and no need to worry about event timing.

It might be worth storing off the event handlers locally to ensure that you don't lose them partway through the process, but I didn't go that far.