Create test case for Flushing.
After review of refactoring, we see that there is a small time which a Flush() might return immediately after a write.

This occurs because the m_TxEmpty event is only cleared when the write has started, not when data is written to an intermediate buffer. The workaround for now is to wait a short period of time (indeterminate, depending on CPU load, e.g. 10-100ms) after the write before flushing.