Reset the parity bit in DCB instead of the Parity byte when there's a difference.
Fixes: https://github.com/jcurl/SerialPortStream/issues/1

When opening the serial port with a parity other than None, it would be reset to None. This was caused by first GetCommState would return the fParity bit set to zero, when one was expected. The GetPortSettings would see a discrepancy as this bit wasn't set so assuming there is no parity and reset it back to None with the setPort flag. By ignoring this bit, we'd get the desired behaviour, that the parity is really Odd. To make the logic complete, we then rely on the Parity byte and set the flag accordingly which works as expected.