Fix reading characters for System.Text.Decoder bug found on Windows also.
After following up from https://bugzilla.xamarin.com/show_bug.cgi?id=40002 and testing on more platforms, this problem has been in Windows since the Dawn of Time (at least since .NET 4.0 on Windows XP). So the bug isn't specific to Mono only and we need to work around it.

Instead of hard coding the bug, let's hope that one day it's fixed. So there's a singleton class that checks for the presence of the bug at run-time. If it's there, then we use the workaround code, else we use the normal code.