When closing/disposing during an active write, we now raise an exception.
This also fixes a some race conditions. On Windows, when closing we need to ensure it is serialised with IsOpen, so that IsOpen occurs afterwards. Otherwise sometimes during the Write operation, we'd exit out of WaitForWrite, but still think we're open. A better way still is to handle this a different way, so that WaitForWrite knows that writes may never occur (as it's not open). WaitForWrite returns on Close in Windows because the CommOverlappedIo thread calls Purge(), which empties the buffer allowing a write to occur, breaking out of the write block.

The unix implementation also serialises Close() and IsOpen, but the test cases haven't been run here yet.