Implement detection of device errors when reading data for Unix.
In addition, we remove the debug output on the console which was left
behind.

We implement a new state, DeviceDead, that is persistent. In this state,
the only thing that is allowed is to close the serial port. All read
operations should behave as if there was a timeout in the first case,
except for ReadTo, which raises an IOException immediately.