Assume the unix library name direct.
We now link directly to 'libnserial.so.1' which implies libnserial.so
with API version 1. The assumption is that you should have the library
for the correct architecture installed in the operating system, or in
the same directory as your executable. If that's not possible, then
write your own dll.config file that is suitable for Mono.

The software is written that there will be no nserial.dll, the windows
implementation uses Windows API direct.