Don't interpret all values in WMI.
Unit test case now passes that compared the serial ports in the system with the
results of the WMI query system on Windows. In a case where a modem might be
installed twice (as on the machine where it was failing), but there is only one
instance active, GetPortDescriptions would return both instances, when only one
instance is actually correct. This means we don't return a COM port that isn't
actually in the system (even if it is registered).