OverlappedResult should wait when completing.
In some cases when testing on Windows 10 using the COM0COM project, we would notice the error 996 being returned, resulting in a read that stopped and no more reading from the serial port. It would be expected that the bWait parameter be irrelevant given that we're acting on a semaphore signal that indicates it *should* be finished, but that appears not to be the case.

In particular, MSDN says, If this parameter is TRUE, and the Internal member of the lpOverlapped structure is STATUS_PENDING, the function does not return until the operation has been completed. If this parameter is FALSE and the operation is still pending, the function returns FALSE and the GetLastError function returns ERROR_IO_INCOMPLETE.

Retested in the original project that was causing the issue and confirmed to be solved. The test case was sending 256kb in one direction using two instances of the SerialPortStream with hardware flow control enabled and the Com0Com library CNCA0 <-> CNCB0. It would occur in that scenario about 4 out of 5 times.