Use a smarter algorithm to get the available ports on Linux
We iterate over /sys/class/tty/* looking for tty's. We check that there's a device associated to ignore psuedotty's and that they must have a kernel major:minor number. For platform:8250 we open the device and check further. Only devices that exist in /dev, and are user accessible which appear real are shown (multiple nodes to the same device are ignored). The description returned is the name of the device given by udev in the uevent file.

The libnserial library is updated to version 1.1.0 as we now have a new C-API. The SerialPortStream remains compatible with libnserial 1.0.x, and if that is installed on the system, SerialPortStream will revert to old behaviour.

Tested on Linux. Simulated v1.0.0 library that fallback works (with EntryPointNotFoundException).

Issue: DOTNET-34