Close a serial port properly on Linux with Native Lib.
In some cases after closing the serial port and immediately trying to reopen it, we'd return EACCES indicating that the serial port is in use. This occurs because of the usage of the ioctl(TIOCEXCL). It's not possible at all anymore to open the serial port. So just before we now close the serial port, we reset the exclusive flag with ioctl(TIOCNXCL), which allows the serial port to always be reopened.

Tested by modifying the Mono test case to run 100 times, which now passes (it didn't before).

Issue: DOTNET-82