Abort pin monitoring on Linux for devices that don't support it
For some serial devices, such as ttyACM* USB devices, the ioctl(TIOCMIWAIT) returns an error. Now if an error occurs, we propogate it to the .NET implementation which already correctly ends the monitoring thread.

Previously, the error was ignored in a loop resulting in 100% CPU utilization.

Issue: DOTNET-85, https://github.com/jcurl/SerialPortStream/issues/22