Fix race condition when checking/aborting modem events.
The method serial_waitformodemevent should only be called at most once. There is already some code to protect the structure used for waiting for the event, but none when cancelling, or when waiting for the event twice (meaning that the code was incomplete and did nothing). This commit finishes the implementation.

In addition, we now report errors when creating the critical section, and errors when creating the cancellable thread.

Issue: DOTNET-86, DOTNET-87