Treat EINTR as a non-fatal interrupt
If a signal occurs for whatever reason, the OS may return EINTR as an error code. This was being treated as a fatal error in the .NET implementation. libnserial now returns zero for serial_read and serial_write as is the case for EAGAIN / EWOULDBLOCK.

The .NET code was optimized slightly so that in case zero is returned, it skips a buffer check and tries again (this change is non-functional).

Issue: DOTNET-91