Make usage of pthread_setcancel{state|type} portable
According to the linux manpage, the second parameter may be NULL, but this is not standard in the Posix specification. We should provide a pointer to an int, even if we don't use it.

Added test cases to ensure that the function serial_waitformodemevent can be aborted by serial_abortwaitformodemevent. Tested that this works. We tested also what happens if the type is PTHREAD_CANCEL_DEFERRED instead of PTHREAD_CANCEL_ASYNCHRONOUS and it turns out that the call to ioctl(TIOCMIWAIT, ...) is not a cancellation point.

Issue: DOTNET-93