Updgrade project to work with .NET Core 1.0.4
Originally provided as a pull request from Danny <dannyb2100@hotmail.com> in https://github.com/jcurl/SerialPortStream/pull/28. Moved the csproj for netstandard 1.5 in the same folder as the rest, provided a solution file for VS2017. The dependencies in the CSPROJ file are now 4.3.0 to match that in the NuSpec file for .NET STandard 1.5.

Build script and NuSpec was updated for the new binaries.

Issue: DOTNET-104, https://github.com/jcurl/SerialPortStream/issues/20