SerialPortStream: The IsDisposable flag should not be publicly settable
Additionally, clean up Dispose() to reduce indenting, and properly do a GC.SuppressFinalize. The GC change is to meet the Dispose programming pattern and is required in case the classes are derived from. That these classes aren't derived from means that this change is cosmetic and a coding standard change only.

All unit tests pass on Windows.

Issue: DOTNET-154