Avoid deadlock in event handling
When an event occurs, a "m_EventLock" was taken to check the number of bytes received which took the lock "m_ReadLock". Separately, the main I/O thread would take the lock "m_ReadLock" and may indicate that an error occurred trying to take "m_EventLock". This may result in a deadlock situation of a read event occurs simultaneously with an error event.

Issue: DOTNET-156, https://github.com/jcurl/SerialPortStream/issues/50