Reduce number of objects created when Writing, Reading, Flushing
Instead of creating a new array of WaitHandle objects on everytime a read, write or flush is done, prepare the arrays in the constructor and use that.

Looking into the .NET 4.7.1 reference sources, the method WaitHandles.WaitAny makes a copy of that array too, so there is no need to create it twice and burden GC0.

Issue: DOTNET-162