Reduce CPU and Memory if tracing is disabled
Before printing out a trace, first check if it would be traced with the TraceSource.ShouldTrace command. Particularly, to make this as fast as possible and have the possibility for inlining, an internal implementation is made caching what trace events may be traced in a bit field.

The GC is improved as we no longer need to create arrays for the TraceSource.TraceEvent(xxx, params object[] args), where the compiler will allocate an array automatically, and then box any value types into the array thus creating more objects on the GC heap.

Issue: DOTNET-163