Don't use Enum.HasFlag
The Enum.HasFlag operation results in two boxing operations and is significantly slower than using integer boolean arithmetic. The boxing arises through converting the Enum type to a reference type, so that HasFlag can be called. The second boxing operation is converting the flag to test from an enum type to a reference type for the test.

This change would also slightly reduce CPU load.

Issue: DOTNET-165