Allow RTS/DTR to be set before serial port is open on UNIX
On Windows, one can already set the RTS and DTR flags. They are cached before the serial port is opened. On Unix, an exception would be raised, due to the serial port not being opened.

Allow consistency between Windows and Unix. The libnserial library implements this caching, so it must be compiled and redeployed.

Issue: https://github.com/jcurl/SerialPortStream/issues/62, DOTNET-168