Windows: Fallback to checking registry if port is not CHAR/UNKNOWN
Some devices, e.g. a uBlox is incorrectly registered in Windows as a file device. This prevents the port from being opened. In this case, check as a last resort that the port exists in the registry. If so, then open it anyway.

Issue: https://github.com/jcurl/SerialPortStream/issues/64, DOTNET-169