SerialPortStream: Don't lose data on write
In the case that the SerialPortStream write buffer is nearly full and the method SerialPortStreamWrite() is called, only a portion of the buffer may be written, with the end of the buffered data being discarded.

The problem was the method SerialBuffer.WaitForWrite would return true even if the requested number of bytes weren't available. The fix is trivial, loop back and wait again after each write.

The problem is most likely to manifest itself when the Windows Serial Driver is likely to return with less data written than requested with the WriteFile API.

Issue: DOTNET-176