SerialPortStream: Fix buffer handling
If a SerialPortStream.Write() would occur, where the input offset was non-zero and the length was such that the internal ring buffer needed to split the copy into two, the second half of the copy would be from the wrong offset resulting in corrupted data when sending.

The problem would occur on writing data out, not receiving, as receiving data would set the OS to write to the empty portions of the array, never causing a split of data.

Issue: DOTNET-183
Issue: https://github.com/jcurl/SerialPortStream/issues/104