CircularBuffer: Check length before decoding with System.Text
Trying to decode a zero length source buffer on Mono may result in an ArgumentException (observed with ISO-8859-15, but not with UTF8/ASCII).

The exception raised also has no parameter set, thus a check needs to be done to ensure that the ParamName is not null, else this situation will result in a double fault with a NullReferenceException also.

Issue: DOTNET-194