SerialPortStream: Correct P/Invoke Signature for ReadFile
The ReadFile() method had the buffer prefixed with [Out] which might mean that the wrong buffer is being received and result in data corruption.

See https://github.com/jcurl/SerialPortStream/issues/110 for a discussion, but it was not confirmed that this is the fault, and was only observed in .NET 3.5 64-bit (which is unsupported by this library). This code won't affect usage on Linux.

Issue: https://github.com/jcurl/SerialPortStream/issues/110