Provide a Docker File for a build environment
Use Docker to provide a build environment instead of having to set up a virtual machine for each target.

Changes to 'build.sh':
- Updated to allow specification of folders (use defaults, but using a docker needs a specific path).
- Spaces in paths should be properly handled
- Minor performance update using ${PWD} instead of `pwd`.

See `dll/docker/README.md` for instructions on building the docker container, and then creating and starting it, to build libnserial. It's much quicker now to test changes to the sources that all operating system versions (including future ones) build without warnings/errors, usually with gcc changes.

The same command can be used to build the debian packages for Ubuntu. As Ubuntu has moved to a single 64-bit architecture, docker containers are a convenient way to produce packages using Docker (previously, 32-bit and 64-bit virtual machines were set up, files were copied across, built, copied back for both architectures and signed). See the `README.md` in this commit for more details on how to run. One can create the docker using the master branch, and then check out to the branch with the debian files and build. As the docker files are generic, it can be used for multiple builds.

Issue: DOTNET-320