libnserial docker: Add a shell and query packages
Add the ability to enter to a shell to perform operations on the container. This can help with debugging too.

Add the ability to query some specific installed packages, used by libnserial build scripts.

The command to query the packages (not in the README, as it's less important, except for debugging scripts):

$ docker run -it --rm --read-only --cap-drop all \
   -v ${PWD}:/source:ro \
   -v ${PWD}/../build:/build:rw \
   --tmpfs /tmp \
   -u $(id -u ${USER}):$(id -g ${USER}) \
   libnserial:${CODENAME} dpkg-query

Issue: DOTNET-324