SerialPortStream Upgrade to .NET Core Project Style for .NET 4.0 and 4.5 (.NET Core still to do)
Migrate the project structure to the newer .NET SDK project style. This allows for future integration with .NET Core and .NET 5.0 and later. The project targets multiple frameworks, including also .NET 4.0 and .NET 4.5.

See BUILD.md for instructions on using `dotnet` command to build. Required to read this for the `Signed_Build` configuration. This also contains information about building on Linux as well as some known errors.

The `.csproj` files are customised per hand, so that project files are opt-in, like in old style Visual Studio projects, instead of opt-out. So if new files are added, they must be explicitly mentioned in the `.csproj` file. The SerialPortStream-netstandard15.csproj will be removed when that is migrated to.

Some Test Cases were fixed after running on Linux:
- ReadAndWriteAsync() didn't read data from the serial port properly, if data was split in hardware buffers. On Linux, it's more likely that only a portion of the data is read.
- ReadAndWriteAsync() was failing due to the previous test Flush() that left data in hardware buffers
- SendReceiveBoundaries() used an incorrect baudrate so failed on Linux

Tested:
- Built and runs with Visual Studio 2019 (16.9)
- Runs on the command line with `dotnet build -c [Debug|Release|Signed_Release]` for Windows 10 and Ubuntu 20.04 with .NET 5.0
- Unit Test Cases run on Windows 10 and Linux with `dotnet test -c [Debug|Release]`
- NuGet Package file is generated, similar attributes to older versions

Issue: DOTNET-329