SerialPortStream: Merge .NET Standard with .csproj
The `SerialPortStream-netstandard15.csproj` file is now removed and part of `SerialPortSTream.csproj`. The conditional files in `System` are now moved to the root namespace and are conditionally included only for .NET Standard 1.5. .NET Standard 1.5 doesn't require the package Mono.Posix,v4.0.

Update the sources with the conditional include from `NETSTANDARD15` to what MSDN uses which is `NETSTANDARD1_5`.

Test cases are not part of this commit, as it is intended to be a rafactoring only. As the previous commits didn't support this, neither will this. The result is that there should be no discernable differences between the old version and this version.

Issue: DOTNET-331