SerialPortStream: Enable logging for .NET Standard 1.5 and later
The SerialPortStream now provides a constructor that the user can give an ILogger object (only on .NET Core). This will be used for logging.

For those using .NET Framework, and moving to .NET Core, it could be a lot of work to rewrite to use the new constructor (as it's unlikely this will be used in a Dependency Injection framework). There's a new static property LogSourceFactory.LoggerFactory, for which an ILoggerFactory can be assigned at the beginning of the program. By default, this is null, and no logging should take place, the way it is today. Otherwise, give it your favourite logger.

On .NET Core, this library will not initialize or read any configuration files for logging, it's up to the user to provide the logging framework, as is the principle for .NET Core logging.

Issue: DOTNET-328