SerialPortStream: Remove the SignedRelease
The signed release allows for the project to be installed in the GAC. So this commit removes that feature, and a further commit is required later to bring this back in.

Removing the Signed_Release is required to allow the next commits to reference other non-signed libraries. References for `InternalsVisibleTo` must always be present.

Issue: DOTNET-406