SerialPortStream: Use a different implementation for IAsyncResult
Use the AsyncResult implementation from RJCP.DLL.BufferIO to handle asynchronous operations. This implementation is slightly more generic and is already in use with other projects for some time.

This also allows us to make the handling for .NET 4.0, 4.5 and later, and .NET Core the same by processing in the background, and letting the AsyncResult implementation do most of the work.

Issue: DOTNET-400