SerialPortStream: Fix Logger injection
Bump to version 2.4.0.

On .NET Core, the user can pass in an ILogger object. If this is done for two instances of SerialPortStream, the second instance is used for logging both.

This commit fixes that the ILogger object is a member of the SerialPortStream, and it is passed to classes as it's instantiated, instead of being assigned to a static property.

Issue: DOTNET-419