SerialPortStream: Refactor internal buffer handling
The implementation of SerialBuffer has been replaced by a new implementation from RJCP.IO.Buffer MemoryReadBuffer (SerialReadBuffer) and MemoryWriteBuffer (SerialWriteBuffer. The code is more modular and easier to test for correctness.

The ReadToCache is now part of SerialReadBuffer, which has access to the underlying buffers in MemoryReadBuffer. This puts the code where it more logically belongs

Logging of ReadToCache is now removed as they were excessive. Logging should not be a substitute for proper unit test cases, which may come later.

Issue: DOTNET-416