SerialPortStream: Add an integration test that checks BytesToWrite
When reviewing code, it was observed that when writing, the buffer for the number of bytes to write is only decremented when the write is returned. If the overlapped operation only completed when the actual write was complete, this could result in BytesToWrite counting the same bytes being sent twice, thus providing a value higher than what reality is.

Tested on various chipsets, and confirmed that when writing, the BytesToWrite would return more data than what is actually written.
* PL2303 RA
* CP2101

Issue: DOTNET-425