SerialPortStream: BytesToWrite returns maximum of driver or internal buffers, not the sum
On Windows, the internal Overlapped I/O when writing may only return when the driver itself has actually sent the number of bytes to the hardware. This results in bytes being added twice.

Thus, we only provide the number of bytes from the driver or the internal buffer, whichever is larger.

Issue: DOTNET-425