SerialPortStream: Move Native Serial implementation to the Serial namespace
This is an internal refactoring. Separate the unmanaged (P/Invoke) code from the managed implementations. The unmanaged and marshalling code should be in "Native", while the managed should be somewhere else.

Later, it is planned that the implementations that monitor the serial ports in managed code be made public (exposed) for better modularity. This code is moved to the Serial namespace.

Issue: DOTNET-404