SerialPortStream: Expose INativeSerial
Expose the implementations of INativeSerial so that these can be used when instantiating the serial port. The native implementations are not intended for general use.

Because the INativeSerial implementation is responsible for also getting the list of available ports, the methods GetPortNames() and GetPortDescriptions() are no longer static. An implementation must first instantiate the SerialPortStream to get this information.

Issue: DOTNET-404