SerialPortStreamTest: Refactor test cases
The SerialPortStream test cases are a combination of unit testing and integration tests. The unit tests, integration tests, manual tests should be separated for a better overview of what tests exist, and to make it easier to implement new unit tests.

Any test that requires a physical serial port (e.g. with a NULL-modem cable) is moved to SerialPortStreamNativeTest. Test cases that need a serial port should not be anywhere else, as the names of serial ports are global and therefore can't run in parallel. Having them all in one test case ensures they don't run in parallel. To not run these tests in the IDE, use the category filter:
 -Trait:"ManualTest"

Manual tests that require physical intervention is moved to SerialPortStreamManualTest. These test cases must be marked as Explicit, so they don't run automatically.

All tests that don't use a serial port should be put in SerialPortStreamTest.

Some minor messages were fixed and suppressed:
* Suppress CA1835:Prefer the 'Memory'-based overloads for 'ReadAsync' and 'WriteAsync'
* Suppress IDE0056:Indexing can be simplified
* Fix CA1822:Member does not access instance data and can be marked as static

Issue: DOTNET-403