SerialPortStream: Convert LibNSerial to be static
As the LibNSerial doesn't have any state, converted all methods to be static.
* CA1822: Member does not access instance data and can be marked as static

This allows to remove the instance creation of m_Dll. The original design decision was to allow the LibNSerial to abstract the library. But the actual implementation did only marshalling between managed and non-managed. With the exposure of INativeSerial, it would be possible to inject the correct flavour for different library implementations.

Issue: DOTNET-435