SerialPortStream: Allow to configure the native driver for Windows
Provide a WinSerialPortStream object, that has a Settings property to set the COMMTIMEOUTS fields when the serial port is opened.

The implementation is more generic than just the Windows serial port, with a little bit of work, this can be extended to any INativeSerial implementation. The driver should implement INativeSerial<T> (which is what WinNativeSerial does), that exposes an additional Settings property. When instantiating, this can be given to the constructor for SerialPortStream<T> which exposes the `T Settings` property, where on Windows this is the `IWinNativeSettings`.

The `WinSerialPortStream` is just a short-hand for `SerialPortStream<IWinNativeSettings>(new WinNativeSerial())`.

A factory method is created, that can check the platform identifier, and create this automatically, depending on the platform. It always returns an `SerialPortStream`, but the type can be tested to see if it implements a specific implementation, where the user can then configure the settings depending on the platform (or use defaults if it doesn't know about the platform, or specifically, if it doesn't handle the specific type that is returned).

This solution provides an extensible way to configure the serial port stream, and is type-safe as it uses Generics.

Issue: DOTNET-410, https://github.com/jcurl/SerialPortStream/issues/74