SerialPortStream: BeginRead didn't check the device state
The `Read()` method would check the device state if it was previously removed and result in an exception, but `BeginRead()` didn't do this, which could result in slightly different behaviour. This behaviour would apply to `ReadAsync()` as being derived from `BeginRead()`.

Now an exception may be raised on `EndRead()`, or the call is done synchronously.

Note, that `ReadAsync()` would raise the exception once the task is started, but one shouldn't rely on this behaviour, it may be raised in the future earlier (to save resources).

Issue: DOTNET-441