SerialPortStream: Fix failing ReadToOverflow
The test case would occasionally fail, as it looks for the string EOF. It would fail, as there is a 11% probability that this string is generated randomly from A-Z characters in a 2048 byte buffer (2045 / 26^3). Run the test 100 times, and it occurs now pretty much always.

Replace the string with something that can't be randomly generated, and that is the lower case "eof".

Issue: DOTNET-442