SerialPortStream: Fix memory leak (buffer) when disposing
On Windows, the buffer is pinned for the overlapped I/O. This buffer wasn't being unpinned, which happens in the memory buffer dispose routine. When a test case is run 1000 times, this could lead to an OutOfMemoryException.

This was introduced when buffer handling was moved from SerialPortStream to the INativeSerial object (in v2.x it was disposed, so memory was freed) in commit 6311ebe ("SerialPortStream: Expose INativeSerial").

Issue: DOTNET-448, DOTNET-404