SerialPortStream: Document issues in running test cases on Linux
Observed that many test cases under SerialPortStreamNativeTest would fail occasionally (about 1%) on Linux. Many of these test cases are done using the SerialPortStream.Virtual and do not fail, indicating a problem when talking to the hardware. It turns out that there are driver specific issues that cannot be resolved using the SerialPortStream. Document these in the README.md file.

Issue: DOTNET-447