SerialPortStream: Add CopyTo and CopyToAsync
Ensure that CopyTo and CopyToAsync work as expected. There is no implementation for these provided by SerialPortStream, and it relies on the base class. To make sure that the base classes are implemented, this was tested on Windows 10 21H1, and Windows 7 x64 SP1 with .NET 4.5.1 installed.

Issue: DOTNET-450, https://github.com/jcurl/RJCP.DLL.SerialPortStream/issues/102