SerialPortStream: Prevent dispose, or exception in event ending program
While testing on Linux, once got a NullReferenceException when accessing the ReadBuffer. This can occur because a modem event may occur before the buffer has been initialized on Linux (they're two different threads, and it depends on which one is initialized first). An exception in the handler will cause the program to crash on .NET Core, as the exception is not handled on a ThreadPool thread.

As on .NET Core, the thread can cause the program to crash, capture all exceptions, as there might be user code outside of the scope of the SerialPortStream that causes this error. The user would probably be confused and may have difficulties identifying the root cause. So allow the program to continue, and stop processing events, and add a log.

Issue: DOTNET-463