SerialPortStream Test: Improve reliability of tests with Com0Com
When running the test case `ReadAndWriteAsyncInDifferentThreadsAtSameTime` multiple times, the first time would pass and the second time would fail when using the Com0Com driver. Using a FTDI with a NULL modem adapter showed no issues.

Enabling logging for Com0Com, we see that the first time the data is written (3 bytes). But the second instance the test runs, it fails with a timeout.

Good Case:
```text
CNCA0/FP c0cWrite WRITE length=3
CNCA0/FP FdoPortWrite WRITE 3: 01 02 03 * ... *, status=SUCCESS
```

Bad Case:
```text
CNCA0/FP c0cWrite WRITE length=3
...
*2* CNCA0/FP complete WRITE 0: *  *, status=TIMEOUT
```

Reading through the code for Com0Com, this appears to be an issue in the driver, that:
a) The first write works, but needs the remote end opened to read the data
b) The second write fails, because the driver appears to be in the state STATUS_PENDING, and doesn't get reset when the ports are closed.

Implementing a thread to open the remote end when it is not normally opened mitigates this issue that the test case passes all the time.

Ran the test cases on Windows 10 with Com0Com 3.0.0.0 and expected test cases now pass. Run with FTDI and a NULL modem adapter and it continues to work.

Issue: DOTNET-711