SerialPortStream: Get devices directly from Windows Device Manager
Use the CfgMgr32 API to query the list of devices and then get details for the device. We can query the device `PortName` from the registry and map that to the listed devices in the system `SERIALCOMM`. From here we observe that the Description/Friendly name is present, and we can now also query the manufacturer to add more context for descriptions that are unclear (e.g. USB Serial [FTDI]).

By not using WMI, we now enable this functionality for .NET Standard 1.5, thus now fully supporting the feature of descriptions on Windows.

Tested on Windows 10.19043.

Issue: DOTNET-698, https://github.com/jcurl/RJCP.DLL.SerialPortStream/issues/136