SerialPortStream: Separate the Description and Manufacturer
Return the Friendly Name instead of the description, that it is similar to the information presented in the Windows Device Manager. The device description is used if the friendly name happens to be blank.

The manufacturer name is no longer part of the description, but is now added as an explicit field in `PortDescription`, which can be used if so desired. The `PortDescription.ToString()` provides a printable string that can be used if desired.

Issue: DOTNET-699, https://github.com/jcurl/RJCP.DLL.SerialPortStream/issues/136