SerialPortStream Test: Reduce time for threaded tests
Most threaded tests existed at the time real serial port hardware was in use. With recent changes, it uses virtual data buffers to test the algorithms and for race conditions.

Shorten some of the timeouts, and use `Task.Yield()` which tries to allow for threads running in parallel (which can be observed by manually inspecting the logs).

Issue: DOTNET-917