SerialPortStream: Remove the license from the header of the files
Removing the license headers from each file is not intended at all to change the already provided license, in particular, the license is still kept as a reference in the `.csproj` file. It is to reduce maintenance for the details in every file and the date, and instead rely on the project metadata.

Issue: DOTNET-956