SerialPortStream: Refactor to remove empty method
There is a Microsoft implementation of System.IO.Ports.SerialPort, but this is intended only for Windows. It doesn't make sense to have such an implementation for Linux. Use conditional includes to reduce code on Linux.

Backported from 2eaca99 ("SerialPortStream: Refactor to remove empty method") on the 3.x branch.

Issue: DOTNET-974