SerialPortFlush: Test program to write and flush data
Set the default `WriteTimeoutConstant` to be zero, which is the same as System.IO.Ports 9.0.4 when the timeout value is not set. It appears that the PL2303 chipset doesn't wait for the writes to complete when this value is non-zero. Further testing shows the CH340 and CH343 always treat this value as zero. Tested the FTDI and CP2012 chipset and they appear to work as expected.

Further updates are to make compatibility changes similar to how the `System.IO.Ports` initialises:
- Change the path to be `\\?\COM` (used to be `\\.\COM`
- Remove completely the `EV_TXEMPTY` event. This was never checked, and we don't set the mask for this any more.

Because this implementation never uses a timeout for the driver (the timeout is managed by this buffer), the data limit is set to 8192 bytes when sending data.

Issue: DOTNET-1030